+++ inject/inject-core/src/test/scala/com/twitter/inject/Resettable.scala
@@ -1,5 +1,6 @@
 package com.twitter.inject
 
+@deprecated("No replacement.", "2020-07-22")
 trait Resettable {
   def reset(): Unit
 }